requirements.txt
numpy
cvlib
pigpio
gpiozero
RPi.GPIO
pymavlink
tensorflow
TIME-python
ultralytics
opencv-python
opencv-contrib-python